SEQNG-285 Match compiler flags as for the gem project #time 30m

+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/main/scala/edu/gemini/seqexec/web/client/components/NavBar.scala
@@ -76,7 +76,7 @@ object ConnectionState {
 
   case class Props(u: WebSocketConnection)
 
-  def formatTime(delay: Long): String = if (delay < 1000) {
+  def formatTime(delay: Int): String = if (delay < 1000) {
     f"${delay / 1000.0}%.1f"
   } else {
     f"${delay / 1000}%d"
